datacentric/pyproject.toml
[tool.poetry]
name = "datacentric"
version = "0.1.0"
description = ""
authors = ["Stef Ruinard <ruinard@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
vertexai = "^1.71.1"
google-cloud-storage = "^2.19.0"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
fsspec = "^2024.10.0"
gcsfs = "^2024.10.0"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
